Update and reorganize decoder tests
- Update tests to use new ChunkDecoder API (append + decodeChunk)
- Split ChunkDecoderTests into separate file for better organization
- Remove 286 lines of commented-out private method tests
- Update MessageDecoderTests to use new MessageDecoder API
- All 352 tests passing
